Drop koinonospool utils in favour of general vapaee::utils module Also move some utilites from vapaee::dex::utils to vapaee::utils
